chore(Structure): Move some functions from crypto to other helpers
