+++ es/utils/string.js
@@ -55,3 +55,14 @@ export function pascalToSnake (s) {
 export function isHex (str) {
   return !!(str.length % 2 === 0 && str.match(/^[0-9a-f]+$/i))
 }
+
+/**
+ * Check whether a string is valid base-64.
+ * @param {string} str String to validate.
+ * @return {boolean} True if the string is valid base-64, false otherwise.
+ */
+export function isBase64 (str) {
+  if (str.length % 4 > 0 || str.match(/[^0-9a-z+/=]/i)) return false
+  const index = str.indexOf('=')
+  return !!(index === -1 || str.slice(index).match(/={1,2}/))
+}
